rename JogInterface to JogPanel
