+++ src/main/java/com/marginallyclever/robotoverlord/components/robot/robotarm/robotArmInterface/jogpanel/CartesianReportPanel.java
@@ -1,4 +1,4 @@
-package com.marginallyclever.robotoverlord.components.robot.robotarm.robotarminterface.joginterface;
+package com.marginallyclever.robotoverlord.components.robot.robotarm.robotarminterface.jogpanel;
 
 import com.marginallyclever.convenience.log.Log;
